Support `many_to_many` without `:through` annotation
Allow a `many_to_many` mapper relationship to not require the `:through`
parameter to annotate the intermediate relationships.

ORMs like Sequel, for example, will be in charge of performing the joins
and requiring to have select fields modified for these intermediate tables.
Therefore there is no need to track them (or select specific fields from them).

This was basically a feature required for the PraxisMapper-native ORM, which
needs to know all the joins/fields of all actual tables.

Signed-off-by: Josep M. Blanquer <blanquer@gmail.com>